+++ src/StateSmith/Input/PlantUML/antlr4/PlantUMLParser.cs
@@ -40,8 +40,8 @@ public partial class PlantUMLParser : Parser {
 		T__9=10, T__10=11, T__11=12, T__12=13, T__13=14, T__14=15, T__15=16, T__16=17, 
 		T__17=18, T__18=19, T__19=20, T__20=21, T__21=22, T__22=23, T__23=24, 
 		T__24=25, T__25=26, START_UML=27, END_UML=28, HWS=29, LINE_ENDER=30, IDENTIFIER=31, 
-		DIGIT=32, KEPT_BLOCK_COMMENT=33, BLOCK_COMMENT=34, LINE_COMMENT=35, SYMBOLS=36, 
-		SINGLE_QUOTE=37, STRING=38, ENDNOTE=39;
+		DIGIT=32, KEPT_BLOCK_COMMENT=33, BLOCK_COMMENT=34, LINE_COMMENT=35, START_OF_INPUT=36, 
+		SYMBOLS=37, SINGLE_QUOTE=38, STRING=39, ENDNOTE=40;
 	public const int
 		RULE_line_ending_ows = 0, RULE_optional_any_space = 1, RULE_ohs = 2, RULE_start_end_state = 3, 
 		RULE_history_state = 4, RULE_state_id = 5, RULE_vertex = 6, RULE_edge = 7, 
@@ -64,14 +64,15 @@ public partial class PlantUMLParser : Parser {
 		null, "'[*]'", "'['", "'h'", "'H'", "']'", "'->'", "'-'", "':'", "'{'", 
 		"'}'", "'<<'", "'>>'", "'state'", "'State'", "'as'", "'hide empty description'", 
 		"'scale'", "'skin'", "'mainframe'", "'title'", "'skinparam'", "'#'", "'note'", 
-		"'on'", "'link'", "'end'", "'@startuml'", "'@enduml'"
+		"'on'", "'link'", "'end'", "'@startuml'", "'@enduml'", null, null, null, 
+		null, null, null, null, "'\u0001'"
 	};
 	private static readonly string[] _SymbolicNames = {
 		null, null, null, null, null, null, null, null, null, null, null, null, 
 		null, null, null, null, null, null, null, null, null, null, null, null, 
 		null, null, null, "START_UML", "END_UML", "HWS", "LINE_ENDER", "IDENTIFIER", 
-		"DIGIT", "KEPT_BLOCK_COMMENT", "BLOCK_COMMENT", "LINE_COMMENT", "SYMBOLS", 
-		"SINGLE_QUOTE", "STRING", "ENDNOTE"
+		"DIGIT", "KEPT_BLOCK_COMMENT", "BLOCK_COMMENT", "LINE_COMMENT", "START_OF_INPUT", 
+		"SYMBOLS", "SINGLE_QUOTE", "STRING", "ENDNOTE"
 	};
 	public static readonly IVocabulary DefaultVocabulary = new Vocabulary(_LiteralNames, _SymbolicNames);
 
@@ -648,7 +649,7 @@ public partial class PlantUMLParser : Parser {
 					State = 99;
 					ErrorHandler.Sync(this);
 					_la = TokenStream.LA(1);
-					while ((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << T__0) | (1L << T__2) | (1L << T__3) | (1L << T__5) | (1L << T__6) | (1L << T__7) | (1L << T__8) | (1L << T__9) | (1L << T__10) | (1L << T__11) | (1L << T__12) | (1L << T__13) | (1L << T__14) | (1L << T__15) | (1L << T__16) | (1L << T__17) | (1L << T__18) | (1L << T__19) | (1L << T__20) | (1L << T__21) | (1L << T__22) | (1L << T__23) | (1L << T__24) | (1L << T__25) | (1L << START_UML) | (1L << END_UML) | (1L << HWS) | (1L << IDENTIFIER) | (1L << DIGIT) | (1L << KEPT_BLOCK_COMMENT) | (1L << BLOCK_COMMENT) | (1L << LINE_COMMENT) | (1L << SYMBOLS) | (1L << SINGLE_QUOTE) | (1L << STRING) | (1L << ENDNOTE))) != 0)) {
+					while ((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << T__0) | (1L << T__2) | (1L << T__3) | (1L << T__5) | (1L << T__6) | (1L << T__7) | (1L << T__8) | (1L << T__9) | (1L << T__10) | (1L << T__11) | (1L << T__12) | (1L << T__13) | (1L << T__14) | (1L << T__15) | (1L << T__16) | (1L << T__17) | (1L << T__18) | (1L << T__19) | (1L << T__20) | (1L << T__21) | (1L << T__22) | (1L << T__23) | (1L << T__24) | (1L << T__25) | (1L << START_UML) | (1L << END_UML) | (1L << HWS) | (1L << IDENTIFIER) | (1L << DIGIT) | (1L << KEPT_BLOCK_COMMENT) | (1L << BLOCK_COMMENT) | (1L << LINE_COMMENT) | (1L << START_OF_INPUT) | (1L << SYMBOLS) | (1L << SINGLE_QUOTE) | (1L << STRING) | (1L << ENDNOTE))) != 0)) {
 						{
 						{
 						State = 96;
@@ -1839,7 +1840,7 @@ public partial class PlantUMLParser : Parser {
 			State = 267;
 			ErrorHandler.Sync(this);
 			_la = TokenStream.LA(1);
-			while ((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << T__0) | (1L << T__1) | (1L << T__2) | (1L << T__3) | (1L << T__4) | (1L << T__5) | (1L << T__6) | (1L << T__7) | (1L << T__8) | (1L << T__9) | (1L << T__10) | (1L << T__11) | (1L << T__12) | (1L << T__13) | (1L << T__14) | (1L << T__15) | (1L << T__16) | (1L << T__17) | (1L << T__18) | (1L << T__19) | (1L << T__20) | (1L << T__21) | (1L << T__22) | (1L << T__23) | (1L << T__24) | (1L << T__25) | (1L << START_UML) | (1L << END_UML) | (1L << HWS) | (1L << IDENTIFIER) | (1L << DIGIT) | (1L << KEPT_BLOCK_COMMENT) | (1L << BLOCK_COMMENT) | (1L << LINE_COMMENT) | (1L << SYMBOLS) | (1L << SINGLE_QUOTE) | (1L << STRING))) != 0)) {
+			while ((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << T__0) | (1L << T__1) | (1L << T__2) | (1L << T__3) | (1L << T__4) | (1L << T__5) | (1L << T__6) | (1L << T__7) | (1L << T__8) | (1L << T__9) | (1L << T__10) | (1L << T__11) | (1L << T__12) | (1L << T__13) | (1L << T__14) | (1L << T__15) | (1L << T__16) | (1L << T__17) | (1L << T__18) | (1L << T__19) | (1L << T__20) | (1L << T__21) | (1L << T__22) | (1L << T__23) | (1L << T__24) | (1L << T__25) | (1L << START_UML) | (1L << END_UML) | (1L << HWS) | (1L << IDENTIFIER) | (1L << DIGIT) | (1L << KEPT_BLOCK_COMMENT) | (1L << BLOCK_COMMENT) | (1L << LINE_COMMENT) | (1L << START_OF_INPUT) | (1L << SYMBOLS) | (1L << SINGLE_QUOTE) | (1L << STRING))) != 0)) {
 				{
 				{
 				State = 264;
@@ -2522,7 +2523,7 @@ public partial class PlantUMLParser : Parser {
 
 	private static char[] _serializedATN = {
 		'\x3', '\x608B', '\xA72A', '\x8133', '\xB9ED', '\x417C', '\x3BE7', '\x7786', 
-		'\x5964', '\x3', ')', '\x16F', '\x4', '\x2', '\t', '\x2', '\x4', '\x3', 
+		'\x5964', '\x3', '*', '\x16F', '\x4', '\x2', '\t', '\x2', '\x4', '\x3', 
 		'\t', '\x3', '\x4', '\x4', '\t', '\x4', '\x4', '\x5', '\t', '\x5', '\x4', 
 		'\x6', '\t', '\x6', '\x4', '\a', '\t', '\a', '\x4', '\b', '\t', '\b', 
 		'\x4', '\t', '\t', '\t', '\x4', '\n', '\t', '\n', '\x4', '\v', '\t', '\v', 
@@ -2605,7 +2606,7 @@ public partial class PlantUMLParser : Parser {
 		'\x34', '\x36', '\x2', '\t', '\x3', '\x2', '\x1F', ' ', '\x3', '\x2', 
 		'\x5', '\x6', '\x5', '\x2', '\x4', '\x4', '\a', '\a', ' ', ' ', '\x3', 
 		'\x2', '\xF', '\x10', '\x3', '\x2', ' ', ' ', '\x4', '\x2', ' ', ' ', 
-		')', ')', '\x6', '\x2', '\xF', '\x11', '\x13', '\x17', '\x19', '\x1C', 
+		'*', '*', '\x6', '\x2', '\xF', '\x11', '\x13', '\x17', '\x19', '\x1C', 
 		'!', '!', '\x2', '\x18B', '\x2', '\x38', '\x3', '\x2', '\x2', '\x2', '\x4', 
 		'?', '\x3', '\x2', '\x2', '\x2', '\x6', '\x45', '\x3', '\x2', '\x2', '\x2', 
 		'\b', 'H', '\x3', '\x2', '\x2', '\x2', '\n', 'J', '\x3', '\x2', '\x2', 
@@ -2677,7 +2678,7 @@ public partial class PlantUMLParser : Parser {
 		'\x8B', '\x3', '\x2', '\x2', '\x2', '\x8D', '\x8E', '\x3', '\x2', '\x2', 
 		'\x2', '\x8E', '\x8C', '\x3', '\x2', '\x2', '\x2', '\x8E', '\x8F', '\x3', 
 		'\x2', '\x2', '\x2', '\x8F', '\x9C', '\x3', '\x2', '\x2', '\x2', '\x90', 
-		'\x92', '\a', '(', '\x2', '\x2', '\x91', '\x93', '\a', '\x1F', '\x2', 
+		'\x92', '\a', ')', '\x2', '\x2', '\x91', '\x93', '\a', '\x1F', '\x2', 
 		'\x2', '\x92', '\x91', '\x3', '\x2', '\x2', '\x2', '\x93', '\x94', '\x3', 
 		'\x2', '\x2', '\x2', '\x94', '\x92', '\x3', '\x2', '\x2', '\x2', '\x94', 
 		'\x95', '\x3', '\x2', '\x2', '\x2', '\x95', '\x96', '\x3', '\x2', '\x2', 
@@ -2789,7 +2790,7 @@ public partial class PlantUMLParser : Parser {
 		'\x2', '\x126', '\x129', '\x3', '\x2', '\x2', '\x2', '\x127', '\x125', 
 		'\x3', '\x2', '\x2', '\x2', '\x127', '\x128', '\x3', '\x2', '\x2', '\x2', 
 		'\x128', '\x12A', '\x3', '\x2', '\x2', '\x2', '\x129', '\x127', '\x3', 
-		'\x2', '\x2', '\x2', '\x12A', '\x12B', '\a', ')', '\x2', '\x2', '\x12B', 
+		'\x2', '\x2', '\x2', '\x12A', '\x12B', '\a', '*', '\x2', '\x2', '\x12B', 
 		'+', '\x3', '\x2', '\x2', '\x2', '\x12C', '\x12E', '\a', '\x19', '\x2', 
 		'\x2', '\x12D', '\x12F', '\a', '\x1F', '\x2', '\x2', '\x12E', '\x12D', 
 		'\x3', '\x2', '\x2', '\x2', '\x12F', '\x130', '\x3', '\x2', '\x2', '\x2', 
@@ -2805,12 +2806,12 @@ public partial class PlantUMLParser : Parser {
 		'\x3', '\x2', '\x2', '\x2', '\x13D', '\x13B', '\x3', '\x2', '\x2', '\x2', 
 		'\x13D', '\x13E', '\x3', '\x2', '\x2', '\x2', '\x13E', '\x140', '\x3', 
 		'\x2', '\x2', '\x2', '\x13F', '\x13D', '\x3', '\x2', '\x2', '\x2', '\x140', 
-		'\x141', '\a', ')', '\x2', '\x2', '\x141', '-', '\x3', '\x2', '\x2', '\x2', 
+		'\x141', '\a', '*', '\x2', '\x2', '\x141', '-', '\x3', '\x2', '\x2', '\x2', 
 		'\x142', '\x144', '\a', '\x19', '\x2', '\x2', '\x143', '\x145', '\a', 
 		'\x1F', '\x2', '\x2', '\x144', '\x143', '\x3', '\x2', '\x2', '\x2', '\x145', 
 		'\x146', '\x3', '\x2', '\x2', '\x2', '\x146', '\x144', '\x3', '\x2', '\x2', 
 		'\x2', '\x146', '\x147', '\x3', '\x2', '\x2', '\x2', '\x147', '\x148', 
-		'\x3', '\x2', '\x2', '\x2', '\x148', '\x149', '\a', '(', '\x2', '\x2', 
+		'\x3', '\x2', '\x2', '\x2', '\x148', '\x149', '\a', ')', '\x2', '\x2', 
 		'\x149', '\x14A', '\x5', '\x1C', '\xF', '\x2', '\x14A', '/', '\x3', '\x2', 
 		'\x2', '\x2', '\x14B', '\x150', '\x5', '$', '\x13', '\x2', '\x14C', '\x150', 
 		'\x5', ',', '\x17', '\x2', '\x14D', '\x150', '\x5', '*', '\x16', '\x2',